requirements.txt
Flask
pandas
pdfplumber
python-docx
pytesseract
pillow
requests